Added a folder for the instances files.
Added the functionality of reading an instance specific file and
printing all instances files in "instances" folder.

Code was refactored to obey object-oriented paradigm:
	-Private methods were created to hide implementation details, e.g,
Graph class

Variable names were changed so as to make the code more readable.

New function for priting the execution information such as the time
elapsed and chromatic number.

Function "run()" of the Application class was rewritten for the purpose
of clarifying the sequence of steps performed by the program.